Rename share file to shares
